Address PR comments - break up the scripts into folders to segment the functionality (hopefully making it a little easier to follow) - Added private: true - Added package-lock.json
