+++ sandbox-tools/merge-repos/src/github/checkPrExists.ts
@@ -17,7 +17,7 @@
 import * as child_process from "child_process";
 import * as util from "util";
 import { SimpleGit} from "simple-git";
-import { log } from "./utils";
+import { log } from "../support/utils";
 import path = require("path");
 
 const execFile = util.promisify(child_process.execFile);
