+++ sandbox-tools/merge-repos/src/support/abort.ts
