+++ sandbox-tools/merge-repos/src/support/clean.ts
