+++ sandbox-tools/merge-repos/src/support/isIgnoreFolder.ts
