+++ sandbox-tools/merge-repos/src/support/processRepos.ts
